Move nglview import out of molecule module
